test(checks): split split-name-safety tests out of test_project_gate_no_empty_split.py
Round 7's new symlink regression test pushed the file to 313 lines, over
the 300-line bloat-baseline guideline (Stop-hook gate). Extracts the two
_is_safe_split_name tests into their own file, following this repo's
established per-concern test-file split precedent — both files now under
the limit.
